Incorrect configuration in docker-compose.yml
Issue:
After making changes and thus triggering auto rebuild with chokidar,
the website is not updated on localhost.

Cause:
Incorrect volume bindings: the "devserver" is not "seeing" the file
changes in the public folder and will always serve the intial files
present after building the container.

Fix:
"generator" and "devserver" services share volume data.
"devserver" is dependent on the "generator" service.